feat: move functionality for `AppData` into module
