+++ src-tauri/crates/cyanocitta_nostr_tools/src/app_data/mod.rs
@@ -1,0 +1,53 @@
+use std::path::PathBuf;
+
+use crate::*;
+use anyhow::{anyhow, Result};
+use serde::{Deserialize, Serialize};
+
+/// AppData.
+#[derive(Default, Deserialize, Serialize)]
+pub struct AppData {
+    /// Information about this user.
+    pub profiles: Vec<Profile>,
+    /// Current profile index.
+    pub current_profile: usize,
+    /// List of relays.
+    pub relays: Vec<Relay>,
+}
+
+impl AppData {
+    /// Get path to where [`AppData`] is stored.
+    fn get_path() -> Result<PathBuf> {
+        let mut path =
+            dirs::data_local_dir().ok_or_else(|| anyhow!("failed getting local data dir"))?;
+        path.push("cyanocitta.app/data.json");
+
+        Ok(path)
+    }
+
+    /// Load [`AppData`] from path.
+    pub fn load() -> Result<AppData> {
+        let path = Self::get_path()?;
+        let app_data = serde_json::from_slice(&std::fs::read(&path)?)?;
+
+        Ok(app_data)
+    }
+
+    /// Save [`AppData`] to path.
+    pub fn save(&self) -> Result<()> {
+        let path = Self::get_path()?;
+        let mut dir_path = path.clone();
+        dir_path.pop();
+
+        std::fs::create_dir_all(&dir_path)?;
+        std::fs::write(&path, serde_json::to_string(self)?)?;
+
+        Ok(())
+    }
+}
+
+impl Drop for AppData {
+    fn drop(&mut self) {
+        self.save().unwrap();
+    }
+}
