+++ src-tauri/crates/cyanocitta_nostr_tools/src/lib.rs
@@ -1,9 +1,11 @@
+mod app_data;
 mod client;
 mod message;
 mod profile;
 mod relay;
 mod tests;
 
+pub use app_data::*;
 pub use client::*;
 pub use message::*;
 pub use profile::*;
