+++ src-tauri/crates/cyanocitta_nostr_tools/src/profile/mod.rs
@@ -1,4 +1,3 @@
-use anyhow::Result;
 use secp256k1::{rand::rngs::OsRng, PublicKey, Secp256k1, SecretKey};
 use serde::{Deserialize, Serialize};
